refactor: order data package into appropriate sub packages
