+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/api/CollectionAPI.kt
@@ -7,8 +7,8 @@ import tech.thatgravyboat.skyblockapi.utils.extentions.asInt
 import tech.thatgravyboat.skyblockapi.utils.extentions.asLong
 import tech.thatgravyboat.skyblockapi.utils.extentions.asString
 import tech.thatgravyboat.skyblockapi.utils.http.Http
-import tech.thatgravyboat.skyblockpv.data.CollectionCategory
-import tech.thatgravyboat.skyblockpv.data.CollectionEntry
+import tech.thatgravyboat.skyblockpv.data.api.CollectionCategory
+import tech.thatgravyboat.skyblockpv.data.api.CollectionEntry
 
 private const val API_URL = "https://api.hypixel.net/v2/resources/skyblock/collections"
