+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/api/data/SkyBlockProfile.kt
@@ -7,25 +7,20 @@ import tech.thatgravyboat.skyblockapi.utils.extentions.*
 import tech.thatgravyboat.skyblockpv.api.CollectionAPI
 import tech.thatgravyboat.skyblockpv.api.ItemAPI
 import tech.thatgravyboat.skyblockpv.api.SkillAPI
-import tech.thatgravyboat.skyblockpv.data.CfData
-import tech.thatgravyboat.skyblockpv.data.CollectionItem
-import tech.thatgravyboat.skyblockpv.data.Currency
-import tech.thatgravyboat.skyblockpv.data.EssenceData
-import tech.thatgravyboat.skyblockpv.data.RiftData
 import tech.thatgravyboat.skyblockpv.data.SortedEntry.Companion.sortToCollectionsOrder
 import tech.thatgravyboat.skyblockpv.data.SortedEntry.Companion.sortToSkillsOrder
 import tech.thatgravyboat.skyblockpv.data.SortedEntry.Companion.sortToSlayerOrder
-import tech.thatgravyboat.skyblockpv.data.skills.FishData
-import tech.thatgravyboat.skyblockpv.data.skills.Pet
-import tech.thatgravyboat.skyblockpv.data.skills.TrophyFishData
-import tech.thatgravyboat.skyblockpv.data.skills.combat.DungeonData
-import tech.thatgravyboat.skyblockpv.data.skills.combat.MobData
-import tech.thatgravyboat.skyblockpv.data.skills.combat.SlayerTypeData
-import tech.thatgravyboat.skyblockpv.data.skills.farming.FarmingData
-import tech.thatgravyboat.skyblockpv.data.skills.farming.GardenData
-import tech.thatgravyboat.skyblockpv.data.skills.mining.Forge
-import tech.thatgravyboat.skyblockpv.data.skills.mining.GlaciteData
-import tech.thatgravyboat.skyblockpv.data.skills.mining.MiningCore
+import tech.thatgravyboat.skyblockpv.data.api.CfData
+import tech.thatgravyboat.skyblockpv.data.api.CollectionItem
+import tech.thatgravyboat.skyblockpv.data.api.Currency
+import tech.thatgravyboat.skyblockpv.data.api.RiftData
+import tech.thatgravyboat.skyblockpv.data.api.skills.*
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.DungeonData
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.MobData
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.SlayerTypeData
+import tech.thatgravyboat.skyblockpv.data.api.skills.farming.FarmingData
+import tech.thatgravyboat.skyblockpv.data.api.skills.farming.GardenData
+import tech.thatgravyboat.skyblockpv.data.repo.EssenceData
 import tech.thatgravyboat.skyblockpv.utils.ChatUtils
 import tech.thatgravyboat.skyblockpv.utils.getPath
 import java.util.*
