+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/data/api/CollectionData.kt
@@ -1,4 +1,4 @@
-package tech.thatgravyboat.skyblockpv.data
+package tech.thatgravyboat.skyblockpv.data.api
 
 import net.minecraft.world.item.ItemStack
