+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/data/api/RiftData.kt
@@ -1,15 +1,10 @@
-package tech.thatgravyboat.skyblockpv.data
+package tech.thatgravyboat.skyblockpv.data.api
 
 import com.google.gson.JsonObject
-import com.mojang.serialization.Codec
-import com.mojang.serialization.JsonOps
-import com.mojang.serialization.codecs.RecordCodecBuilder
 import net.minecraft.world.item.ItemStack
 import tech.thatgravyboat.skyblockapi.utils.extentions.asInt
 import tech.thatgravyboat.skyblockapi.utils.extentions.asList
-import tech.thatgravyboat.skyblockpv.api.ItemAPI
-import tech.thatgravyboat.skyblockpv.data.skills.Pet
-import tech.thatgravyboat.skyblockpv.utils.Utils
+import tech.thatgravyboat.skyblockpv.data.api.skills.Pet
 import tech.thatgravyboat.skyblockpv.utils.getNbt
 import tech.thatgravyboat.skyblockpv.utils.getPath
 import tech.thatgravyboat.skyblockpv.utils.legacyStack
@@ -94,50 +89,5 @@ data class Trophy(
                 visits = json["visits"].asInt,
             )
         }
-    }
-}
-
-object RiftCodecs {
-    var data: RiftRepoData? = null
-        private set
-
-    private val trophyCodec = RecordCodecBuilder.create {
-        it.group(
-            Codec.STRING.fieldOf("id").forGetter(TrophyRepo::id),
-            Codec.STRING.fieldOf("name").forGetter(TrophyRepo::name),
-        ).apply(it, ::TrophyRepo)
-    }
-
-    init {
-        val CODEC = RecordCodecBuilder.create {
-            it.group(
-                trophyCodec.listOf().fieldOf("trophies").forGetter(RiftRepoData::trophies),
-                Codec.STRING.listOf().fieldOf("montezuma").forGetter(RiftRepoData::montezuma),
-                Codec.STRING.listOf().fieldOf("eyes").forGetter(RiftRepoData::eyes),
-            ).apply(it, ::RiftRepoData)
-        }
-
-        val cfData = Utils.loadFromRepo<JsonObject>("rift") ?: JsonObject()
-
-        CODEC.parse(JsonOps.INSTANCE, cfData).let {
-            if (it.isError) {
-                throw RuntimeException(it.error().get().message())
-            }
-            data = it.getOrThrow()
-        }
-    }
-
-    data class RiftRepoData(
-        val trophies: List<TrophyRepo>,
-        val montezuma: List<String>,
-        val eyes: List<String>,
-    )
-
-    data class TrophyRepo(
-        val id: String,
-        val name: String,
-    ) {
-        val hypixelId = "RIFT_TROPHY_${id.uppercase()}"
-        val item by lazy { ItemAPI.getItem(hypixelId) }
     }
 }
