+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/data/repo/SkullTextures.kt
@@ -1,4 +1,4 @@
-package tech.thatgravyboat.skyblockpv.data
+package tech.thatgravyboat.skyblockpv.data.repo
 
 import tech.thatgravyboat.skyblockpv.utils.Utils
