+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/MainScreen.kt
@@ -22,9 +22,9 @@ import tech.thatgravyboat.skyblockpv.api.SkillAPI.getSkillLevel
 import tech.thatgravyboat.skyblockpv.api.StatusAPI
 import tech.thatgravyboat.skyblockpv.api.data.PlayerStatus
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.SkullTextures
-import tech.thatgravyboat.skyblockpv.data.skills.combat.getIconFromSlayerName
-import tech.thatgravyboat.skyblockpv.data.skills.combat.getSlayerLevel
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.getIconFromSlayerName
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.getSlayerLevel
+import tech.thatgravyboat.skyblockpv.data.repo.SkullTextures
 import tech.thatgravyboat.skyblockpv.screens.BasePvScreen
 import tech.thatgravyboat.skyblockpv.screens.elements.ExtraConstants
 import tech.thatgravyboat.skyblockpv.utils.FakePlayer
