+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/combat/MobScreen.kt
@@ -7,7 +7,7 @@ import net.minecraft.client.gui.layouts.Layout
 import net.minecraft.client.gui.layouts.LinearLayout
 import tech.thatgravyboat.skyblockapi.utils.extentions.toFormattedString
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.skills.combat.MobData
+import tech.thatgravyboat.skyblockpv.data.api.skills.combat.MobData
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.LayoutUtils.centerHorizontally
 import tech.thatgravyboat.skyblockpv.utils.displays.DisplayWidget
