+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/farming/BaseFarmingScreen.kt
@@ -17,7 +17,7 @@ import tech.thatgravyboat.skyblockapi.utils.text.TextStyle.color
 import tech.thatgravyboat.skyblockpv.api.GardenApi
 import tech.thatgravyboat.skyblockpv.api.ItemAPI
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.skills.farming.GardenProfile
+import tech.thatgravyboat.skyblockpv.data.api.skills.farming.GardenProfile
 import tech.thatgravyboat.skyblockpv.screens.tabs.base.AbstractCategorizedScreen
 import tech.thatgravyboat.skyblockpv.screens.tabs.base.Category
 import kotlin.reflect.KClass
