+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/farming/FarmingScreen.kt
@@ -18,8 +18,12 @@ import tech.thatgravyboat.skyblockpv.api.ItemAPI
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
 import tech.thatgravyboat.skyblockpv.api.predicates.ItemPredicateHelper
 import tech.thatgravyboat.skyblockpv.api.predicates.ItemPredicates
-import tech.thatgravyboat.skyblockpv.data.skills.Pet
-import tech.thatgravyboat.skyblockpv.data.skills.farming.*
+import tech.thatgravyboat.skyblockpv.data.api.skills.Pet
+import tech.thatgravyboat.skyblockpv.data.api.skills.farming.FarmingData
+import tech.thatgravyboat.skyblockpv.data.api.skills.farming.MedalType
+import tech.thatgravyboat.skyblockpv.data.repo.FarmingGear
+import tech.thatgravyboat.skyblockpv.data.repo.GardenResource
+import tech.thatgravyboat.skyblockpv.data.repo.StaticGardenData
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.Utils.append
 import tech.thatgravyboat.skyblockpv.utils.Utils.rightPad
