+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/mining/HotmScreen.kt
@@ -16,16 +16,14 @@ import tech.thatgravyboat.skyblockapi.utils.text.TextStyle.bold
 import tech.thatgravyboat.skyblockapi.utils.text.TextStyle.color
 import tech.thatgravyboat.skyblockapi.utils.text.TextStyle.strikethrough
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.skills.mining.*
+import tech.thatgravyboat.skyblockpv.data.api.skills.MiningCore
+import tech.thatgravyboat.skyblockpv.data.api.skills.PowderType
+import tech.thatgravyboat.skyblockpv.data.repo.*
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.LayoutUtils.asScrollable
 import tech.thatgravyboat.skyblockpv.utils.LayoutUtils.withScrollToBottom
 import tech.thatgravyboat.skyblockpv.utils.Utils.round
-import tech.thatgravyboat.skyblockpv.utils.displays.Display
-import tech.thatgravyboat.skyblockpv.utils.displays.DisplayWidget
-import tech.thatgravyboat.skyblockpv.utils.displays.Displays
-import tech.thatgravyboat.skyblockpv.utils.displays.asWidget
-import tech.thatgravyboat.skyblockpv.utils.displays.withTooltip
+import tech.thatgravyboat.skyblockpv.utils.displays.*
 import java.text.DecimalFormat
 import kotlin.math.absoluteValue
 import kotlin.math.roundToInt
