+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/mining/MainMiningScreen.kt
@@ -13,10 +13,10 @@ import tech.thatgravyboat.skyblockpv.SkyBlockPv
 import tech.thatgravyboat.skyblockpv.api.ItemAPI
 import tech.thatgravyboat.skyblockpv.api.RemindersAPI
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.EssenceData.addMiningPerk
-import tech.thatgravyboat.skyblockpv.data.skills.mining.ForgeTimeData
-import tech.thatgravyboat.skyblockpv.data.skills.mining.MiningCore
-import tech.thatgravyboat.skyblockpv.data.skills.mining.RockBrackets
+import tech.thatgravyboat.skyblockpv.data.api.skills.MiningCore
+import tech.thatgravyboat.skyblockpv.data.api.skills.RockBrackets
+import tech.thatgravyboat.skyblockpv.data.repo.EssenceData.addMiningPerk
+import tech.thatgravyboat.skyblockpv.data.repo.ForgeTimeData
 import tech.thatgravyboat.skyblockpv.utils.ChatUtils
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.LayoutUtils.asScrollable
