+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/mining/MiningGearScreen.kt
@@ -7,7 +7,7 @@ import tech.thatgravyboat.skyblockapi.api.datatype.DataType
 import tech.thatgravyboat.skyblockapi.api.datatype.DataTypes
 import tech.thatgravyboat.skyblockapi.api.datatype.getData
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.skills.mining.MiningGear
+import tech.thatgravyboat.skyblockpv.data.api.skills.MiningGear
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.components.PvWidgets
 import tech.thatgravyboat.skyblockpv.utils.displays.DisplayWidget
