+++ src/main/kotlin/tech/thatgravyboat/skyblockpv/screens/tabs/rift/MainRiftScreen.kt
@@ -11,8 +11,8 @@ import tech.thatgravyboat.skyblockapi.utils.text.Text
 import tech.thatgravyboat.skyblockapi.utils.text.TextColor
 import tech.thatgravyboat.skyblockapi.utils.text.TextStyle.color
 import tech.thatgravyboat.skyblockpv.api.data.SkyBlockProfile
-import tech.thatgravyboat.skyblockpv.data.RiftCodecs
-import tech.thatgravyboat.skyblockpv.data.RiftData
+import tech.thatgravyboat.skyblockpv.data.api.RiftData
+import tech.thatgravyboat.skyblockpv.data.repo.RiftCodecs
 import tech.thatgravyboat.skyblockpv.utils.LayoutBuild
 import tech.thatgravyboat.skyblockpv.utils.LayoutUtils.asScrollable
 import tech.thatgravyboat.skyblockpv.utils.Utils.append
